Absolute positioning for Inputs and Mixes list.
